backend/requirements.txt
pyspark==3.2.4
cassandra-driver
django-cassandra-engine